requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dataclasses-json==0.6.7
deepdiff==8.0.1
deepl==1.19.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
emoji==2.14.0
faiss-cpu==1.8.0.post1
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.9.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
jiter==0.5.0
joblib==1.4.2
JPype1==1.5.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwipiepy==0.18.1
kiwipiepy_model==0.18.0
konlpy==0.6.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-teddynote==0.0.34
langchain-text-splitters==0.3.0
langchain-upstage==0.3.0
langdetect==1.0.9
langsmith==0.1.130
llama-cloud==0.1.2
llama-index==0.11.16
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.16
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.11
llama-index-multi-modal-llms-openai==0.2.2
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.7
lxml==5.3.0
lz4==4.3.3
markdownify==0.13.1
marshmallow==3.22.0
mmh3==4.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.51.0
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.10.0
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.15.13
unstructured-client==0.25.9
urllib3==2.2.3
wget==3.2
wrapt==1.16.0
yarl==1.13.1
